+++ packages/container/config/oktaAuthConfig.js
@@ -2,14 +2,14 @@ const oktaAuthConfig = {
     // Note: If your app is configured to use the Implicit Flow
     // instead of the Authorization Code with Proof of Code Key Exchange (PKCE)
     // you will need to add `pkce: false`
-    issuer: 'https://{yourOktaDomain}/oauth2/default',
-    clientId: '{clientId}',
+    issuer: 'https://dev-9755339.okta.com/oauth2/default',
+    clientId: '0oa3asmnoJt6Aj4ji5d6',
     redirectUri: window.location.origin + '/login/callback',
-  };
-  
-  const oktaSignInConfig = {
-    baseUrl: 'https://{yourOktaDomain}',
-    clientId: '{clientId}',
+};
+
+const oktaSignInConfig = {
+    baseUrl: 'https://dev-9755339.okta.com',
+    clientId: '0oa3asmnoJt6Aj4ji5d6',
     redirectUri: window.location.origin + '/login/callback',
     authParams: {
       // If your app is configured to use the Implicit Flow
@@ -18,5 +18,5 @@ const oktaAuthConfig = {
       // pkce: false
     }
   };
-  
-  export { oktaAuthConfig, oktaSignInConfig };
+
+export {oktaAuthConfig , oktaSignInConfig};
